requirements.txt
gpustat
tensorboardX
torchsummary
icecream
tqdm
protobuf==3.20
wandb==0.16.6
